Initial Source Code Upload
This commit includes the initial upload of Graphi(t)'s source code. The uploaded code comprises all the features and functionalities required to create and visualize 3D parametric plots with ease.

Key features included in this upload:

Intuitive interface for defining parametric functions.
Adjustable time range for precise control over the plotted interval.
Dynamic iteration capability for function evaluation.
Support for additional variables for enhanced flexibility in plotting.
Save and load functionality to store and retrieve plots.
This marks the beginning of Graphi(t)'s journey in the open-source community. We look forward to the enhancements and improvements that future commits will bring to this project.